refactor: restructure `Engine` logic
very clunky as of now, but serves as a basis for this refactor
