+++ tetris/__init__.py
@@ -4,7 +4,6 @@ import importlib.metadata
 
 from tetris import engine
 from tetris import types
-from tetris.engine import DefaultEngine
 from tetris.engine import Engine
 from tetris.game import BaseGame
 from tetris.types import MinoType
@@ -19,7 +18,6 @@ from tetris.types import PlayingStatus
 __version__ = importlib.metadata.version("tetris")
 __all__ = (
     "BaseGame",
-    "DefaultEngine",
     "engine",
     "Engine",
     "MinoType",
